separated responsibilities and added features
separated auto loading responsibilities to autoloading class.
Dispatcher.php now only throws the following exceptions:
badClassNameException, classMethodNotFoundException,
classNotSpecifiedException, methodNotSpecifiedException. The auto
loader handles file not found and class not found with
AutoLoaderClassFileNotFoundException and
AutoLoaderClassNotFoundException

Added class name spacing

Allow match string to be a function that accepts the match parameter
and returns true if it is a match or false if it is not